Autenticazione funzionante, db sincronizzati automaticamente

+++ VeraPelle/src/main/java/psw/verapelle/service/CategoryService.java
@@ -7,6 +7,8 @@ import psw.verapelle.DTO.CategoryDTO;
 import psw.verapelle.entity.Category;
 import psw.verapelle.repository.CategoryRepository;
 
+import java.util.Objects;
+
 @Service
 public class CategoryService {
 
@@ -15,6 +17,9 @@ public class CategoryService {
 
     @Transactional
     public Category createCategory(CategoryDTO categoryDTO) {
+        if (categoryDTO.getName() == null || categoryDTO.getName().trim().isEmpty()) {
+            throw new IllegalArgumentException("Category name cannot be empty");
+        }
         Category category = new Category(null, categoryDTO.getName(), categoryDTO.getDescription(), null, null);
         return categoryRepository.save(category);
     }
@@ -23,14 +28,20 @@ public class CategoryService {
     public void deleteCategory(Long categoryId) {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new RuntimeException("Category not found"));
+
+        if (!category.getProducts().isEmpty()) {
+            throw new RuntimeException("Cannot delete category, it has associated products.");
+        }
+
         categoryRepository.delete(category);
     }
+
 
     @Transactional
     public Category updateCategory(Long id, Category updatedCategory) {
         return categoryRepository.findById(id).map(category -> {
-            category.setName(updatedCategory.getName());
-            category.setDescription(updatedCategory.getDescription());
+            category.setName(Objects.requireNonNullElse(updatedCategory.getName(), category.getName())); //ritorna il primo se non nullo altrimenti il secondo se non nullo
+            category.setDescription(Objects.requireNonNullElse(updatedCategory.getDescription(), category.getDescription()));
             return categoryRepository.save(category);
         }).orElseThrow(() -> new RuntimeException("Category not found"));
     }
